% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_jot_user}
\alias{get_jot_user}
\title{Get Current User}
\usage{
get_jot_user(user = "")
}
\arguments{
\item{user}{user to override get with. Default is \code{''}, where it checks
options, then system info.}
}
\value{
character
}
\description{
Get Current User
}
\examples{
get_jot_user()
get_jot_user(NA_character_)
}
\concept{helpers}
