% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{jot_erase}
\alias{jot_erase}
\title{Erase a Note}
\usage{
jot_erase(name, pad = jot_active())
}
\arguments{
\item{name}{character, name of note to erase. required.}

\item{pad}{path to file of pad to write to. default is \code{jot_active()}.}
}
\value{
path to pad, invisibly
}
\description{
Erase a Note
}
\examples{
tf <- tempfile()
jot_new_pad(tf)
jot(3, 'example')
jot_erase('example')
}
\concept{write}
