% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jot_list}
\alias{jot_list}
\title{List out names of values in a jot notepad}
\usage{
jot_list(pad = jot_active())
}
\arguments{
\item{pad}{path to notepad to open}
}
\value{
vlist of names of values in notepad
}
\description{
List out names of values in a jot notepad
}
\examples{
tf <- tempfile()
jot_new_pad(tf)
jot_activate(tf)
jot('ex_val', 'ex')
jot_list()
}
\concept{read}
