% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{jot_lock}
\alias{jot_lock}
\alias{jot_unlock}
\title{Lock or Unlock a jot Notepad}
\usage{
jot_lock(pad = jot_active())

jot_unlock(pad = jot_active())
}
\arguments{
\item{pad}{path to file to open}
}
\value{
path to pad, invisibly
}
\description{
Lock or Unlock a jot Notepad
}
\examples{
tf <- tempfile()
new_pad(tf)
jot_activate(tf)
jot_lock()
jot_unlock()
}
