% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{jot_new_pad}
\alias{jot_new_pad}
\alias{jot_new}
\title{Create new jot Notepad}
\usage{
jot_new_pad(pad = ".jot.yaml", title)

jot_new(pad = ".jot.yaml", title)
}
\arguments{
\item{pad}{path to file to store new notepad.}

\item{title}{what to call the notepad. If not supplied, uses the name of the file.}
}
\value{
path to file, invisibly
}
\description{
Create new jot Notepad
}
\examples{
tf <- tempfile()
jot_new_pad(tf)
}
\concept{pad}
