% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{jot_open}
\alias{jot_open}
\alias{jot_activate}
\alias{jot_active}
\alias{jot_current}
\alias{jot_close}
\alias{jot_deactivate}
\title{Open and Close jot Notepads}
\usage{
jot_open(pad)

jot_activate(pad)

jot_active()

jot_current()

jot_close()

jot_deactivate()
}
\arguments{
\item{pad}{path to file to open}
}
\value{
path to pad, invisibly
}
\description{
Open and Close jot Notepads
}
\examples{
tf <- tempfile()
new_pad(tf)
jot_open(tf)
jot_active()
jot_current()
jot_close()
jot_active()
}
