% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jot_read}
\alias{jot_read}
\alias{read_jot}
\title{Read a jot notepad}
\usage{
jot_read(name, pad = jot_active())

read_jot(name, pad = jot_active())
}
\arguments{
\item{name}{name of noteto read}

\item{pad}{path to file to open}
}
\value{
value specified or all values
}
\description{
Read a jot notepad
}
\examples{
tf <- tempfile()
jot_new_pad(tf)
jot_activate(tf)
jot(note = 'ex_note', name = 'ex_name')
jot_read()
jot_read('ex_name')
}
\concept{read}
