% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{new_pad}
\alias{new_pad}
\alias{new_jot}
\title{Create new jot Notepad}
\usage{
new_pad(file = ".jot.yaml", title)

new_jot(file = ".jot.yaml", title)
}
\arguments{
\item{file}{path to file}

\item{title}{what to call the notepad. If not supplied, uses the name of the file.}
}
\value{
path to file, invisibly
}
\description{
Create new jot Notepad
}
\examples{
tf <- tempfile()
new_pad(tf)
}
