% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jot_read}
\alias{jot_read}
\alias{read_jot}
\title{Read a jot Notepad}
\usage{
jot_read(field, pad = jot_active())

read_jot(field, pad = jot_active())
}
\arguments{
\item{field}{field to read}

\item{pad}{path to file to open}
}
\value{
value specified or all values
}
\description{
Read a jot Notepad
}
\examples{
tf <- tempfile()
new_pad(tf)
jot_activate(tf)
jot(note = 'ex_note', name = 'ex_name')
jot_read()
jot_read('ex_name')
}
