% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jot_skim}
\alias{jot_skim}
\title{Read in all values from a jot Notepad}
\usage{
jot_skim(pad = jot_active())
}
\arguments{
\item{pad}{path to file to open}
}
\value{
value specified or all values
}
\description{
Read in all values from a jot Notepad
}
\examples{
tf <- tempfile()
new_pad(tf)
jot_activate(tf)
jot('ex_val', 'ex')
jot_skim()
}
